forge fmt

+++ src/PriceConverter.sol
@@ -1,66 +1,67 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.19;
 // import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/shared/interfaces/AggregatorV3Interface.sol";
+
 import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 
-library PriceConverter{
-
+library PriceConverter {
     // function getDataFeedEth() internal pure returns (AggregatorV3Interface){
     //     AggregatorV3Interface dataFeed = AggregatorV3Interface(0x694AA1769357215DE4FAC081bf1f309aDC325306);
     //     return dataFeed;
     // }
 
-    function getDataFeedBtc() internal pure returns (AggregatorV3Interface){
+    function getDataFeedBtc() internal pure returns (AggregatorV3Interface) {
         AggregatorV3Interface dataFeedBtcWithEth = AggregatorV3Interface(0x5fb1616F78dA7aFC9FF79e0371741a747D2a7F22);
         return dataFeedBtcWithEth;
     }
 
-
     function getChainlinkDataFeedLatestAnswer(AggregatorV3Interface dataFeed) internal view returns (uint256) {
-
         (
-        /* uint80 roundId */
-        ,
-        int256 answer,
-        /*uint256 startedAt*/
-        ,
-        /*uint256 updatedAt*/
-        ,
-        /*uint80 answeredInRound*/
+            /* uint80 roundId */
+            ,
+            int256 answer,
+            /*uint256 startedAt*/
+            ,
+            /*uint256 updatedAt*/
+            ,
+            /*uint80 answeredInRound*/
         ) = dataFeed.latestRoundData();
         return uint256(answer) * 1e10;
     }
 
-    function getLatestBtcPriceInEth() internal view returns (uint256){
-        AggregatorV3Interface  dataFeedBtcWithEth = getDataFeedBtc();
+    function getLatestBtcPriceInEth() internal view returns (uint256) {
+        AggregatorV3Interface dataFeedBtcWithEth = getDataFeedBtc();
         (, int256 price,,,) = dataFeedBtcWithEth.latestRoundData();
         return uint256(price);
     }
 
-    function getChainlinkDataFeedLatestVersion(AggregatorV3Interface dataFeed) internal view returns (uint256 version){
+    function getChainlinkDataFeedLatestVersion(AggregatorV3Interface dataFeed)
+        internal
+        view
+        returns (uint256 version)
+    {
         version = dataFeed.version();
     }
 
-    function getChainlinkDataFeedLatestdecimals(AggregatorV3Interface dataFeed) internal view returns (uint8){
+    function getChainlinkDataFeedLatestdecimals(AggregatorV3Interface dataFeed) internal view returns (uint8) {
         return dataFeed.decimals();
     }
 
-    function getPrice(AggregatorV3Interface dataFeed) internal view returns(uint256) {
-
-        (,int price,,,) = dataFeed.latestRoundData();
+    function getPrice(AggregatorV3Interface dataFeed) internal view returns (uint256) {
+        (, int256 price,,,) = dataFeed.latestRoundData();
         // int256 price = 2000 * 1e18;
 
-        return uint256(price) * 1E10;
+        return uint256(price) * 1e10;
         // return uint256(price);
     }
 
-    function getConversionRate(uint256 ethAmount, AggregatorV3Interface _dataFeed) internal view returns(uint256) {
+    function getConversionRate(uint256 ethAmount, AggregatorV3Interface _dataFeed) internal view returns (uint256) {
         uint256 ethPrice = getPrice(_dataFeed);
         uint256 ethAmountInUsd = (ethPrice * ethAmount) / 1e18;
         return ethAmountInUsd;
     }
 
-    function conversUstToEth(uint256 usdAmount, AggregatorV3Interface _dataFeed) internal view returns(uint256){
+    function conversUstToEth(uint256 usdAmount, AggregatorV3Interface _dataFeed) internal view returns (uint256) {
         uint256 ethPric = getPrice(_dataFeed);
         uint256 ethAmount = usdAmount * 1e18 / ethPric;
         return ethAmount;
